LayoutTests/http/tests/websocket/tests/hybi/useragent-in-openinghandshake-expected.txt

Web Socket test to check if opening handshake contains User-Agent header

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.data is navigator.userAgent
PASS successfullyParsed is true

TEST COMPLETE
